requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2024.8.30
cffi==1.17.1
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
numpy==2.2.6
openai==1.30.0
packaging==24.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.9
python-dotenv==1.0.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
websocket-client==1.8.0
